+++ CHANGELOG.md
@@ -17,6 +17,7 @@
   * Added Automation preview.
   * Added lifecycle hook for variable readiness.
   * Added datatypes for variables.
+  * Added a feature to render datatype.
 
 ## v0.1.7 (2022-06-28)
   * Created database if it does not exist before migraions.